[#52] Move input component into components directory
