+++ src/routes/account.tsx
@@ -3,7 +3,7 @@ import {getCookie} from 'hono/cookie'
 import Layout from '@/layout'
 import bcrypt from 'bcryptjs'
 import {AUTH_COOKIE_NAME, isLoggedIn} from '@/services/auth'
-import Input from '@/input'
+import Input from '@/components/input'
 import {db} from '@/db'
 import {users} from '@/db/schema'
 import {eq} from 'drizzle-orm'
