+++ src/routes/panel.tsx
@@ -2,7 +2,7 @@ import {Hono} from 'hono'
 import {AUTH_COOKIE_NAME, isLoggedIn} from '@/services/auth'
 import {getCookie} from 'hono/cookie'
 import Layout from '@/layout'
-import Input from '@/input'
+import Input from '@/components/input'
 import {getWebUntis} from '@/services/untis'
 import {randomUUID} from 'node:crypto'
 import {db} from '@/db'
